+++ packages/react-contexts/src/middlewares/index.ts
@@ -1,12 +1,17 @@
 import { chain } from './chain'
 import type { MiddlewareFactory } from './types'
-import { refreshSessionMiddleware } from './refresh-session'
 import { setWebDeviceIdMiddleware } from './set-web-device-id'
+import { serverRefreshSessionMiddleware } from './session/server'
 
 export const constructMiddleware = (functions: MiddlewareFactory[]) =>
-  chain([...functions, refreshSessionMiddleware, setWebDeviceIdMiddleware])
+  chain([
+    ...functions,
+    serverRefreshSessionMiddleware,
+    setWebDeviceIdMiddleware,
+  ])
 
 export * from './types'
 export { chain } from './chain'
-export { refreshSessionMiddleware } from './refresh-session'
+export { serverRefreshSessionMiddleware } from './session/server'
 export { setWebDeviceIdMiddleware } from './set-web-device-id'
+export { applySetCookie } from './utils/apply-set-cookie'
