+++ packages/react-contexts/src/middlewares/session/server.ts
@@ -1,23 +1,21 @@
-import { type ResponseCookie } from 'next/dist/compiled/@edge-runtime/cookies'
 import {
   NextFetchEvent,
   NextMiddleware,
   NextRequest,
   NextResponse,
 } from 'next/server'
-import { get, post } from '@titicaca/fetcher'
-import { parseString, splitCookiesString } from 'set-cookie-parser'
+import { get } from '@titicaca/fetcher'
 import {
   TP_SE,
   TP_TK,
   SESSION_KEY as X_SOTO_SESSION,
 } from '@titicaca/constants'
 
-import { parseApp } from '../user-agent-context'
+import { parseApp } from '../../user-agent-context'
 
-import { applySetCookie } from './utils/apply-set-cookie'
+import { getSessionRefreshResponse } from './refresh-session'
 
-export function refreshSessionMiddleware(next: NextMiddleware) {
+export function serverRefreshSessionMiddleware(next: NextMiddleware) {
   return async function middleware(
     request: NextRequest,
     event: NextFetchEvent,
@@ -59,30 +57,14 @@ export function refreshSessionMiddleware(next: NextMiddleware) {
       return response
     }
 
-    /**
-     * /web-session/token은 TP-TK의 유효성을 확인해서 TP_TK, TP_SE, x-soto-session 응답합니다.
-     */
-    const refreshResponse = await post('/api/users/web-session/token', options)
+    const sessionRefreshResponse = await getSessionRefreshResponse({
+      next,
+      request,
+      event,
+      options,
+    })
 
-    if (refreshResponse.ok) {
-      const setCookie = refreshResponse.headers.get('set-cookie')
-
-      if (setCookie) {
-        const response = (await next(request, event)) as NextResponse
-        const setCookies = splitCookiesString(setCookie)
-        setCookies.forEach((cookie) => {
-          const { name, value, ...rest } = parseString(cookie)
-          if (name !== X_SOTO_SESSION) {
-            response.cookies.set(name, value, { ...(rest as ResponseCookie) })
-          }
-        })
-        applySetCookie(request, response)
-
-        return response
-      }
-    }
-
-    return response
+    return sessionRefreshResponse || response
   }
 }
